Fixed state transition bugs

+++ Assets/ScriptableObjects/Runtime/DungeonGeneration/Spawners/SpawnerData.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 public class SpawnerData : ScriptableObject
 {
     public GameObject itemToSpawn;
-    public int minSpawnRate, maxSpawnRate;
+    public float minSpawnRate, maxSpawnRate;
 }
